Fix Team context menu entries not showing in certain editors
Apparently WTP does things differently than other editors. The default
selection object is not meaningful there, so extend the enablement
expression to check activeEditorInput as well.

Also, when executing the command, the selection object is both an
ITextSelection and an empty IStructuredSelection. By checking for
ITextSelection first, we get a result instead of an empty selection.

Bug: 449308
Change-Id: I703a4bd19811b6434dee92228ef3441091d37c9d
Signed-off-by: Robin Stocker <robin@nibor.org>